pyproject.toml
[tool.poetry]
name = "jukebox"
version = "0.2.1"
description = "An Open Source Music Streaming App."
authors = ["JukeBox Developers <arpan.mahanty.007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.101.1"
uvicorn = "^0.23.2"
rocketry = "^2.5.1"
pydantic = "^1.10.12"   # don't update to v2 until all dependents are updated properly
email-validator = "^2.0.0.post2"
yoyo-migrations = "^8.2.0"
psycopg2-binary = "^2.9.7"
google-auth = "^2.22.0"
psutil = "^5.9.5"
pyyaml = "^6.0.1"
sqlmodel = "^0.0.8"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
watchfiles = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
